+++ src/Action/Identity/IdentityOverview.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Identity;
 
 use Doctrine\DBAL\Connection;
+use Heptacom\HeptaConnect\Dataset\Base\UnsafeClassString;
 use Heptacom\HeptaConnect\Storage\Base\Action\Identity\Overview\IdentityOverviewCriteria;
 use Heptacom\HeptaConnect\Storage\Base\Action\Identity\Overview\IdentityOverviewResult;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityOverviewActionInterface;
@@ -38,7 +39,7 @@ final class IdentityOverview implements IdentityOverviewActionInterface
         $externalIdFilter = $criteria->getExternalIdFilter();
         $portalNodeKeyFilter = $criteria->getPortalNodeKeyFilter();
 
-        if ($mappingNodeKeyFilter->count() > 0) {
+        if (!$mappingNodeKeyFilter->isEmpty()) {
             $mappingNodeIds = [];
 
             foreach ($mappingNodeKeyFilter as $mappingNodeKey) {
@@ -55,7 +56,7 @@ final class IdentityOverview implements IdentityOverviewActionInterface
 
         if ($entityTypeFilter !== []) {
             $builder->andWhere($builder->expr()->in('entity_type.type', ':entityTypes'));
-            $builder->setParameter('entityTypes', $entityTypeFilter, Connection::PARAM_STR_ARRAY);
+            $builder->setParameter('entityTypes', \array_map('strval', $entityTypeFilter), Connection::PARAM_STR_ARRAY);
         }
 
         if ($externalIdFilter !== []) {
@@ -63,7 +64,7 @@ final class IdentityOverview implements IdentityOverviewActionInterface
             $builder->setParameter('externalIds', $externalIdFilter, Connection::PARAM_STR_ARRAY);
         }
 
-        if ($portalNodeKeyFilter->count() > 0) {
+        if (!$portalNodeKeyFilter->isEmpty()) {
             $portalNodeIds = [];
 
             foreach ($portalNodeKeyFilter as $portalNodeKey) {
@@ -134,14 +135,14 @@ final class IdentityOverview implements IdentityOverviewActionInterface
                 new PortalNodeStorageKey(Id::toHex((string) $row['portal_node_id'])),
                 new MappingNodeStorageKey(Id::toHex((string) $row['mapping_node_id'])),
                 (string) $row['mapping_external_id'],
-                (string) $row['entity_type_type'],
+                new UnsafeClassString((string) $row['entity_type_type']),
                 /* @phpstan-ignore-next-line */
                 DateTime::fromStorage((string) $row['created_at'])
             )
         );
     }
 
-    protected function getBuilderCached(): QueryBuilder
+    private function getBuilderCached(): QueryBuilder
     {
         if (!$this->builder instanceof QueryBuilder) {
             $this->builder = $this->getBuilder();
@@ -153,7 +154,7 @@ final class IdentityOverview implements IdentityOverviewActionInterface
         return clone $this->builder;
     }
 
-    protected function getBuilder(): QueryBuilder
+    private function getBuilder(): QueryBuilder
     {
         $builder = $this->queryFactory->createBuilder(self::OVERVIEW_QUERY);