+++ src/Action/Job/JobGet.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Job;
 
 use Doctrine\DBAL\Connection;
+use Heptacom\HeptaConnect\Dataset\Base\EntityType;
 use Heptacom\HeptaConnect\Portal\Base\Mapping\MappingComponentStruct;
 use Heptacom\HeptaConnect\Storage\Base\Action\Job\Get\JobGetCriteria;
 use Heptacom\HeptaConnect\Storage\Base\Action\Job\Get\JobGetResult;
@@ -58,7 +59,7 @@ final class JobGet implements JobGetActionInterface
         return $ids === [] ? [] : $this->yieldJobs($ids);
     }
 
-    protected function getBuilderCached(): QueryBuilder
+    private function getBuilderCached(): QueryBuilder
     {
         if (!$this->builder instanceof QueryBuilder) {
             $this->builder = $this->getBuilder();
@@ -70,7 +71,7 @@ final class JobGet implements JobGetActionInterface
         return clone $this->builder;
     }
 
-    protected function getBuilder(): QueryBuilder
+    private function getBuilder(): QueryBuilder
     {
         $builder = $this->queryFactory->createBuilder(self::FETCH_QUERY);
 
@@ -118,7 +119,7 @@ final class JobGet implements JobGetActionInterface
      *
      * @return iterable<JobGetResult>
      */
-    protected function yieldJobs(array $ids): iterable
+    private function yieldJobs(array $ids): iterable
     {
         $builder = $this->getBuilderCached();
         $builder->setParameter('ids', Id::toBinaryList($ids), Connection::PARAM_STR_ARRAY);
@@ -130,7 +131,7 @@ final class JobGet implements JobGetActionInterface
                 new JobStorageKey(Id::toHex((string) $row['job_id'])),
                 new MappingComponentStruct(
                     new PortalNodeStorageKey(Id::toHex((string) $row['portal_node_id'])),
-                    (string) $row['job_entity_type'],
+                    new EntityType((string) $row['job_entity_type']),
                     (string) $row['job_external_id']
                 ),
                 $this->unserializePayload($row['job_payload_payload'], (string) $row['job_payload_format'])