+++ src/Action/PortalExtension/PortalExtensionActivate.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalExtension;
 
+use Heptacom\HeptaConnect\Portal\Base\Portal\PortalExtensionType;
 use Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Activate\PortalExtensionActivatePayload;
 use Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Activate\PortalExtensionActivateResult;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalExtension\PortalExtensionActivateActionInterface;
@@ -15,7 +16,9 @@ final class PortalExtensionActivate extends PortalExtensionSwitchActive implemen
         $payloadExtensions = $payload->getExtensions();
 
         $pass = $this->toggle($payload->getPortalNodeKey()->withoutAlias(), $payloadExtensions);
-        $fail = \array_diff($payloadExtensions, $pass);
+        $fail = $payloadExtensions->filter(
+            static fn (PortalExtensionType $type): bool => !$pass->contains($type)
+        );
 
         return new PortalExtensionActivateResult($pass, $fail);
     }