+++ src/Action/PortalExtension/PortalExtensionDeactivate.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalExtension;
 
+use Heptacom\HeptaConnect\Portal\Base\Portal\PortalExtensionType;
 use Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Deactivate\PortalExtensionDeactivatePayload;
 use Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Deactivate\PortalExtensionDeactivateResult;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalExtension\PortalExtensionDeactivateActionInterface;
@@ -15,7 +16,9 @@ final class PortalExtensionDeactivate extends PortalExtensionSwitchActive implem
         $payloadExtensions = $payload->getExtensions();
 
         $pass = $this->toggle($payload->getPortalNodeKey()->withoutAlias(), $payloadExtensions);
-        $fail = \array_diff($payloadExtensions, $pass);
+        $fail = $payloadExtensions->filter(
+            static fn (PortalExtensionType $type): bool => !$pass->contains($type)
+        );
 
         return new PortalExtensionDeactivateResult($pass, $fail);
     }