+++ src/Action/Route/RouteFind.php
@@ -46,7 +46,7 @@ final class RouteFind implements RouteFindActionInterface
 
         $builder->setParameter('source_key', Id::toBinary($sourceKey->getUuid()), ParameterType::BINARY);
         $builder->setParameter('target_key', Id::toBinary($targetKey->getUuid()), ParameterType::BINARY);
-        $builder->setParameter('type', $criteria->getEntityType());
+        $builder->setParameter('type', (string) $criteria->getEntityType());
 
         $id = $builder->fetchSingleValue();
 
@@ -57,7 +57,7 @@ final class RouteFind implements RouteFindActionInterface
         return new RouteFindResult(new RouteStorageKey(Id::toHex($id)));
     }
 
-    protected function getBuilderCached(): QueryBuilder
+    private function getBuilderCached(): QueryBuilder
     {
         if (!$this->builder instanceof QueryBuilder) {
             $this->builder = $this->getBuilder();
@@ -69,7 +69,7 @@ final class RouteFind implements RouteFindActionInterface
         return clone $this->builder;
     }
 
-    protected function getBuilder(): QueryBuilder
+    private function getBuilder(): QueryBuilder
     {
         $builder = $this->queryFactory->createBuilder(self::LOOKUP_QUERY);