+++ src/Action/RouteCapability/RouteCapabilityOverview.php
@@ -75,7 +75,7 @@ final class RouteCapabilityOverview implements RouteCapabilityOverviewActionInte
         );
     }
 
-    protected function getBuilderCached(): QueryBuilder
+    private function getBuilderCached(): QueryBuilder
     {
         if (!$this->builder instanceof QueryBuilder) {
             $this->builder = $this->getBuilder();
@@ -87,7 +87,7 @@ final class RouteCapabilityOverview implements RouteCapabilityOverviewActionInte
         return clone $this->builder;
     }
 
-    protected function getBuilder(): QueryBuilder
+    private function getBuilder(): QueryBuilder
     {
         $builder = $this->queryFactory->createBuilder(self::OVERVIEW_QUERY);