+++ test/Action/IdentityPersistTest.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Storage\ShopwareDal\Test\Action;
 
 use Doctrine\DBAL\Types\Types;
-use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalContract;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\MappingNodeKeyInterface;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\MappingNodeKeyCollection;
@@ -33,6 +32,7 @@ use Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Id;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Query\QueryFactory;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Query\QueryIterator;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Test\Fixture\Dataset\Simple;
+use Heptacom\HeptaConnect\Storage\ShopwareDal\Test\Fixture\Portal\Portal;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Test\TestCase;
 
 /**
@@ -41,8 +41,10 @@ use Heptacom\HeptaConnect\Storage\ShopwareDal\Test\TestCase;
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalNode\PortalNodeCreate
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Bridge\StorageFacade
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\EntityTypeAccessor
+ * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\PortalNodeAliasAccessor
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKeyGenerator
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey\AbstractStorageKey
+ * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey\PortalNodeStorageKey
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Support\DateTime
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Id
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Query\QueryBuilder
@@ -76,8 +78,8 @@ class IdentityPersistTest extends TestCase
     public function testMergingMappingNodes(): void
     {
         $portalNodeCreateResult = $this->portalNodeCreateAction->create(new PortalNodeCreatePayloads([
-            new PortalNodeCreatePayload(PortalContract::class),
-            new PortalNodeCreatePayload(PortalContract::class),
+            new PortalNodeCreatePayload(Portal::class()),
+            new PortalNodeCreatePayload(Portal::class()),
         ]));
 
         $externalIdSource = Id::randomHex();
@@ -106,8 +108,8 @@ class IdentityPersistTest extends TestCase
     public function testMergingMappingNodesAfterTheSourceChangedPrimaryKeyOrTargetMapsObjectTwice(): void
     {
         $portalNodeCreateResult = $this->portalNodeCreateAction->create(new PortalNodeCreatePayloads([
-            new PortalNodeCreatePayload(PortalContract::class),
-            new PortalNodeCreatePayload(PortalContract::class),
+            new PortalNodeCreatePayload(Portal::class()),
+            new PortalNodeCreatePayload(Portal::class()),
         ]));
 
         $portalNodeKeySource = $portalNodeCreateResult[0]->getPortalNodeKey();
@@ -147,8 +149,8 @@ class IdentityPersistTest extends TestCase
         $externalIdTarget = Id::randomHex();
 
         $portalNodeCreateResult = $this->portalNodeCreateAction->create(new PortalNodeCreatePayloads([
-            new PortalNodeCreatePayload(PortalContract::class),
-            new PortalNodeCreatePayload(PortalContract::class),
+            new PortalNodeCreatePayload(Portal::class()),
+            new PortalNodeCreatePayload(Portal::class()),
         ]));
         $portalNodeKeySource = $portalNodeCreateResult[0]->getPortalNodeKey();
         $portalNodeKeyTarget = $portalNodeCreateResult[1]->getPortalNodeKey();
@@ -172,7 +174,7 @@ class IdentityPersistTest extends TestCase
         static::assertTrue($targetIdentity->getMappingNodeKey()->equals($mappingNodeKey));
         static::assertTrue($targetIdentity->getPortalNodeKey()->equals($portalNodeKeyTarget));
         static::assertSame($externalIdTarget, $targetIdentity->getExternalId());
-        static::assertSame(Simple::class, $targetIdentity->getEntityType());
+        static::assertTrue(Simple::class()->equals($targetIdentity->getEntityType()));
     }
 
     public function testPersistingSameExternalIdToTwoDifferentMappingNodes(): void
@@ -182,8 +184,8 @@ class IdentityPersistTest extends TestCase
         $externalIdTarget = Id::randomHex();
 
         $portalNodeCreateResult = $this->portalNodeCreateAction->create(new PortalNodeCreatePayloads([
-            new PortalNodeCreatePayload(PortalContract::class),
-            new PortalNodeCreatePayload(PortalContract::class),
+            new PortalNodeCreatePayload(Portal::class()),
+            new PortalNodeCreatePayload(Portal::class()),
         ]));
         $portalNodeKeySource = $portalNodeCreateResult[0]->getPortalNodeKey();
         $portalNodeKeyTarget = $portalNodeCreateResult[1]->getPortalNodeKey();
@@ -221,8 +223,8 @@ class IdentityPersistTest extends TestCase
         $externalId2Target = Id::randomHex();
 
         $portalNodeCreateResult = $this->portalNodeCreateAction->create(new PortalNodeCreatePayloads([
-            new PortalNodeCreatePayload(PortalContract::class),
-            new PortalNodeCreatePayload(PortalContract::class),
+            new PortalNodeCreatePayload(Portal::class()),
+            new PortalNodeCreatePayload(Portal::class()),
         ]));
         $portalNodeKeySource = $portalNodeCreateResult[0]->getPortalNodeKey();
         $portalNodeKeyTarget = $portalNodeCreateResult[1]->getPortalNodeKey();
@@ -257,8 +259,8 @@ class IdentityPersistTest extends TestCase
         $externalId2Target = Id::randomHex();
 
         $portalNodeCreateResult = $this->portalNodeCreateAction->create(new PortalNodeCreatePayloads([
-            new PortalNodeCreatePayload(PortalContract::class),
-            new PortalNodeCreatePayload(PortalContract::class),
+            new PortalNodeCreatePayload(Portal::class()),
+            new PortalNodeCreatePayload(Portal::class()),
         ]));
         $portalNodeKeySource = $portalNodeCreateResult[0]->getPortalNodeKey();
         $portalNodeKeyTarget = $portalNodeCreateResult[1]->getPortalNodeKey();
@@ -295,8 +297,8 @@ class IdentityPersistTest extends TestCase
         $externalIdTargetB = Id::randomHex();
 
         $portalNodeCreateResult = $this->portalNodeCreateAction->create(new PortalNodeCreatePayloads([
-            new PortalNodeCreatePayload(PortalContract::class),
-            new PortalNodeCreatePayload(PortalContract::class),
+            new PortalNodeCreatePayload(Portal::class()),
+            new PortalNodeCreatePayload(Portal::class()),
         ]));
         $portalNodeKeySource = $portalNodeCreateResult[0]->getPortalNodeKey();
         $portalNodeKeyTarget = $portalNodeCreateResult[1]->getPortalNodeKey();
@@ -325,12 +327,12 @@ class IdentityPersistTest extends TestCase
         static::assertTrue($targetMappingA->getMappingNodeKey()->equals($mappingNodeKeyA));
         static::assertTrue($targetMappingA->getPortalNodeKey()->equals($portalNodeKeyTarget));
         static::assertSame($externalIdTargetB, $targetMappingA->getExternalId());
-        static::assertSame(Simple::class, $targetMappingA->getEntityType());
+        static::assertTrue(Simple::class()->equals($targetMappingA->getEntityType()));
 
         static::assertTrue($targetMappingB->getMappingNodeKey()->equals($mappingNodeKeyB));
         static::assertTrue($targetMappingB->getPortalNodeKey()->equals($portalNodeKeyTarget));
         static::assertSame($externalIdTargetA, $targetMappingB->getExternalId());
-        static::assertSame(Simple::class, $targetMappingB->getEntityType());
+        static::assertTrue(Simple::class()->equals($targetMappingB->getEntityType()));
     }
 
     public function testDeletingMappingNode(): void
@@ -339,8 +341,8 @@ class IdentityPersistTest extends TestCase
         $externalIdTarget = Id::randomHex();
 
         $portalNodeCreateResult = $this->portalNodeCreateAction->create(new PortalNodeCreatePayloads([
-            new PortalNodeCreatePayload(PortalContract::class),
-            new PortalNodeCreatePayload(PortalContract::class),
+            new PortalNodeCreatePayload(Portal::class()),
+            new PortalNodeCreatePayload(Portal::class()),
         ]));
         $portalNodeKeySource = $portalNodeCreateResult[0]->getPortalNodeKey();
         $portalNodeKeyTarget = $portalNodeCreateResult[1]->getPortalNodeKey();
@@ -376,7 +378,7 @@ class IdentityPersistTest extends TestCase
         $externalIdSource = Id::randomHex();
 
         $portalNodeCreateResult = $this->portalNodeCreateAction->create(new PortalNodeCreatePayloads([
-            new PortalNodeCreatePayload(PortalContract::class),
+            new PortalNodeCreatePayload(Portal::class()),
         ]));
         $portalNodeKeySource = $portalNodeCreateResult[0]->getPortalNodeKey();