+++ test/StorageKeyGeneratorTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Storage\ShopwareDal\Test;
 
-use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalContract;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\IdentityErrorKeyInterface;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\MappingNodeKeyInterface;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
@@ -16,6 +15,7 @@ use Heptacom\HeptaConnect\Storage\Base\Exception\UnsupportedStorageKeyException;
 use Heptacom\HeptaConnect\Storage\Base\PreviewPortalNodeKey;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Bridge\StorageFacade;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey\AbstractStorageKey;
+use Heptacom\HeptaConnect\Storage\ShopwareDal\Test\Fixture\Portal\Portal;
 
 /**
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Bridge\StorageFacade
@@ -48,19 +48,19 @@ class StorageKeyGeneratorTest extends TestCase
     public function testPreviewKeySerialization(): void
     {
         $generator = $this->createStorageFacade()->getStorageKeyGenerator();
-        $serialized = $generator->serialize(new PreviewPortalNodeKey(PortalContract::class));
+        $serialized = $generator->serialize(new PreviewPortalNodeKey(Portal::class()));
 
-        static::assertStringContainsString(\addcslashes(PortalContract::class, '\\'), $serialized);
+        static::assertStringContainsString(\addcslashes(Portal::class, '\\'), $serialized);
     }
 
     public function testPreviewKeyDeserialization(): void
     {
         $generator = $this->createStorageFacade()->getStorageKeyGenerator();
-        $deserialized = $generator->deserialize('{"preview":"Heptacom\\\\HeptaConnect\\\\Portal\\\\Base\\\\Portal\\\\Contract\\\\PortalContract"}');
+        $deserialized = $generator->deserialize('{"preview":"Heptacom\\\\HeptaConnect\\\\Storage\\\\ShopwareDal\\\\Test\\\\Fixture\\\\Portal\\\\Portal"}');
 
         static::assertInstanceOf(PreviewPortalNodeKey::class, $deserialized);
         /* @var $deserialized PreviewPortalNodeKey */
-        static::assertSame(PortalContract::class, $deserialized->getPortalType());
+        static::assertTrue(Portal::class()->equals($deserialized->getPortalType()));
     }
 
     /**