+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The version numbers are structured like `GENERATION.MAJOR.MINOR.PATCH`:
 
 ### Changed
 
+- Change `$payloadExtensions` parameter in `\Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalExtension\PortalExtensionSwitchActive::toggle` to be `\Heptacom\HeptaConnect\Portal\Base\Portal\PortalExtensionTypeCollection` instead of an array of string for better [type safe class strings](https://heptaconnect.io/reference/adr/2022-06-12-type-safe-class-strings/)
+- Change return type of `\Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalExtension\PortalExtensionSwitchActive::toggle` to be a `\Heptacom\HeptaConnect\Portal\Base\Portal\PortalExtensionTypeCollection` instead of an array of string for better [type safe class strings](https://heptaconnect.io/reference/adr/2022-06-12-type-safe-class-strings/)
+
 ### Deprecated
 
 ### Removed
@@ -24,6 +27,25 @@ The version numbers are structured like `GENERATION.MAJOR.MINOR.PATCH`:
 ### Fixed
 
 ### Security
+
+## [0.9.0.3] - 2022-09-20
+
+### Fixed
+
+- Fix error when creating mappings via `\Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Identity\IdentityReflect`. The insert-query now uses the correct table-name.
+
+## [0.9.0.2] - 2022-07-12
+
+### Fixed
+
+- Fix error when deleting many jobs at once by chunking job deletion to 1000 jobs at a time in `Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Job\JobDelete::delete`
+- Fix issue in validation before mapping-node merging involving deleted mappings in `\Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Identity\IdentityPersist::validateMappingNodesCanBeMerged`
+
+## [0.9.0.1] - 2022-04-19
+
+### Fixed
+
+- Fix error related to foreign key checks in migration `\Heptacom\HeptaConnect\Storage\ShopwareDal\Migration\Migration1639860447UpdateExistingJobData`
 
 ## [0.9.0.0] - 2022-04-02